+++ src/components/hero/components/hero-footer.js
@@ -1,16 +1,23 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import classnames from 'classnames';
 
 import Element from '../../element';
 
-const HeroFooter = ({ children, className, ...props }) => (
-  <Element {...props} className={classnames(className, 'hero-foot')}>
-    {children}
-  </Element>
+const HeroFooter = ({ className, ...props }) => (
+  <Element {...props} className={classnames(className, 'hero-foot')} />
 );
 
 HeroFooter.propTypes = {
-  ...Element.propTypes,
+  renderAs: PropTypes.oneOfType([
+    PropTypes.func,
+    PropTypes.string,
+    PropTypes.object,
+  ]),
+};
+
+HeroFooter.defaultProps = {
+  renderAs: 'footer',
 };
 
 export default HeroFooter;