+++ src/components/hero/hero.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import CONSTANTS from '../../constants';
 
-import HeroHead from './components/hero-head';
+import HeroHeader from './components/hero-header';
 import HeroBody from './components/hero-body';
 import HeroFooter from './components/hero-footer';
 
@@ -33,22 +33,25 @@ const Hero = ({
   </Element>
 );
 
-Hero.Head = HeroHead;
+Hero.Header = HeroHeader;
 
 Hero.Body = HeroBody;
 
 Hero.Footer = HeroFooter;
 
 Hero.propTypes = {
-  ...Element.propTypes,
   color: PropTypes.oneOf(colors),
   gradient: PropTypes.bool,
   size: PropTypes.oneOf(['small', 'medium', 'large', 'fullheight']),
   hasNavbar: PropTypes.bool,
+  renderAs: PropTypes.oneOfType([
+    PropTypes.func,
+    PropTypes.string,
+    PropTypes.object,
+  ]),
 };
 
 Hero.defaultProps = {
-  ...Element.defaultProps,
   renderAs: 'section',
 };