+++ src/components/hero/hero.story.js
@@ -1,113 +1,42 @@
 import React from 'react';
 
-import { storiesOf } from '@storybook/react';
+import { Hero, Heading, Container, Notification } from '../..';
+import CONSTANTS from '../../constants';
 
-import { Hero, Heading, Section, Container } from '../..';
-
-storiesOf('Hero', module)
-  .add('Default with color', () => (
-    <div>
-      <Section>
-        <Hero color="primary">
-          <Hero.Body>
-            <Container>
-              <Heading>Hero title Primary</Heading>
-              <Heading subtitle size={3}>
-                Subtitle
-              </Heading>
-            </Container>
-          </Hero.Body>
-        </Hero>
-      </Section>
-
-      <Section>
-        <Hero color="danger">
-          <Hero.Body>
-            <Container>
-              <Heading>Hero title Danger</Heading>
-              <Heading subtitle size={3}>
-                Subtitle
-              </Heading>
-            </Container>
-          </Hero.Body>
-        </Hero>
-      </Section>
-      <Section>
-        <Hero color="info">
-          <Hero.Body>
-            <Container>
-              <Heading>Hero title Info</Heading>
-              <Heading subtitle size={3}>
-                Subtitle
-              </Heading>
-            </Container>
-          </Hero.Body>
-        </Hero>
-      </Section>
-    </div>
-  ))
-  .add('Gradient (EXPERIMENTAL)', () => (
-    <div>
-      <Section>
-        <Hero color="primary" gradient>
-          <Hero.Body>
-            <Container>
-              <Heading>Hero title Primary</Heading>
-              <Heading subtitle size={3}>
-                Subtitle
-              </Heading>
-            </Container>
-          </Hero.Body>
-        </Hero>
-      </Section>
-
-      <Section>
-        <Hero color="danger" gradient>
-          <Hero.Body>
-            <Container>
-              <Heading>Hero title Danger</Heading>
-              <Heading subtitle size={3}>
-                Subtitle
-              </Heading>
-            </Container>
-          </Hero.Body>
-        </Hero>
-      </Section>
-      <Section>
-        <Hero color="info" gradient>
-          <Hero.Body>
-            <Container>
-              <Heading>Hero title Info</Heading>
-              <Heading subtitle size={3}>
-                Subtitle
-              </Heading>
-            </Container>
-          </Hero.Body>
-        </Hero>
-      </Section>
-    </div>
-  ))
-  .add('Differents Sizes', () => (
-    <div>
-      <Hero color="primary" size="medium">
-        <Hero.Body>Medium</Hero.Body>
-      </Hero>
-      <Hero color="info" size="large">
-        <Hero.Body>Large</Hero.Body>
-      </Hero>
-      <Hero color="danger" size="fullheight">
-        <Hero.Body>FULL HEIGHT</Hero.Body>
+export const Default = (args) => {
+  return (
+    <div style={{ margin: '-1rem' }}>
+      <Hero {...args}>
+        <Hero.Header renderAs="header">
+          <Notification color="dark">Header</Notification>
+        </Hero.Header>
+        <Hero.Body>
+          <Container>
+            <Heading>Hero title Info</Heading>
+            <Heading subtitle size={3}>
+              Subtitle
+            </Heading>
+          </Container>
+        </Hero.Body>
+        <Hero.Footer>
+          <Notification color="dark">Footer</Notification>
+        </Hero.Footer>
       </Hero>
     </div>
-  ))
-  .add('Vertical Alignments', () => (
-    <Hero size="fullheight" color="primary">
-      <Hero.Head renderAs="header">
-        <div className="bd-notification is-info">Header</div>
-      </Hero.Head>
-      <Hero.Body>Body</Hero.Body>
-      <Hero.Footer>
-        <div className="bd-notification is-danger">Footer</div>
-      </Hero.Footer>
-    </Hero>
-  ));
+  );
+};
+
+Default.argTypes = {
+  color: {
+    control: {
+      type: 'select',
+      options: ['default'].concat(Object.values(CONSTANTS.COLORS)),
+    },
+  },
+  size: {
+    control: {
+      type: 'select',
+      options: ['default', 'small', 'medium', 'large', 'fullheight'],
+    },
+  },
+};