+++ src/components/hero/index.d.ts
@@ -9,7 +9,7 @@ interface HeroProps {
 }
 
 declare const Hero: BulmaComponent<HeroProps, 'section'> & {
-  Head: BulmaComponent<{}, 'div'>;
+  Header: BulmaComponent<{}, 'div'>;
   Body: BulmaComponent<{}, 'div'>;
   Footer: BulmaComponent<{}, 'div'>;
 };